# docs/openapi/usuarios-service-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SCLI - Usuarios API",
    "version": "1.0.0",
    "description": "Snapshot determinista derivado de controladores y DTO del código fuente vigente."
  },
  "paths": {
...
    "/api/v1/docentes/perfil/{perfilId}": {
      "get": {
        "operationId": "DocenteController_obtenerPorPerfilId",
        "tags": [
          "DocenteController"
        ],
        "responses": {
          "200": {
            "description": "Respuesta satisfactoria",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocenteResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "perfilId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v1/docentes/{id}": {
      "get": {
        "operationId": "DocenteController_obtenerPorId",
        "tags": [
          "DocenteController"
        ],
        "responses": {
          "200": {
            "description": "Respuesta satisfactoria",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocenteResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "operationId": "DocenteController_actualizar",
        "tags": [
          "DocenteController"
        ],
        "responses": {
          "200": {
            "description": "Respuesta satisfactoria",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocenteResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocenteRequest"
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v1/internal/docentes/perfil/{perfilId}": {
      "get": {
        "operationId": "InternalDocenteController_obtenerPorPerfilId",
        "tags": [
          "InternalDocenteController"
        ],
        "responses": {
          "200": {
            "description": "Respuesta satisfactoria",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocenteInternoResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "perfilId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "security": [
          {
            "internalApiKey": []
          }
        ]
      }
    },
    "/api/v1/internal/docentes/{docenteId}": {
      "get": {
        "operationId": "InternalDocenteController_obtenerPorId",
        "tags": [
          "InternalDocenteController"
        ],
        "responses": {
          "200": {
            "description": "Respuesta satisfactoria",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocenteInternoResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "docenteId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "security": [
          {
            "internalApiKey": []
          }
        ]
      }
    },
    "/api/v1/internal/docentes/{docenteId}/carreras/{carreraId}/exists": {
      "get": {
        "operationId": "InternalDocenteController_perteneceCarrera",
        "tags": [
          "InternalDocenteController"
        ],
        "responses": {
          "200": {
            "description": "Respuesta satisfactoria",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "docenteId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "carreraId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "security": [
          {
            "internalApiKey": []
          }
        ]
      }
    },
...
    "/api/v1/perfiles/administracion-usuarios": {
      "post": {
        "operationId": "UsuarioInstitucionalAdminController_crear",
        "tags": [
          "UsuarioInstitucionalAdminController"
        ],
        "responses": {
          "200": {
            "description": "Respuesta satisfactoria",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PerfilResponse"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UsuarioInstitucionalCreateRequest"
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/perfiles/administracion-usuarios/{perfilId}": {
      "put": {
        "operationId": "UsuarioInstitucionalAdminController_actualizar",
        "tags": [
          "UsuarioInstitucionalAdminController"
        ],
        "responses": {
          "200": {
            "description": "Respuesta satisfactoria",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PerfilResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "perfilId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UsuarioInstitucionalUpdateRequest"
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "internalApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Internal-Api-Key"
      }
    },
    "schemas": {
...
      "DocenteInternoResponse": {
        "type": "object",
        "properties": {
          "docenteId": {
            "type": "string",
            "format": "uuid"
          },
          "perfilId": {
            "type": "string",
            "format": "uuid"
          },
          "activo": {
            "type": "boolean"
          }
        }
      },
      "DocenteRequest": {
        "type": "object",
        "properties": {
          "perfilId": {
            "type": "string",
            "format": "uuid"
          },
          "codigoDocente": {
            "type": "string"
          },
          "tituloAcademico": {
            "type": "string"
          },
          "departamento": {
            "type": "string"
          },
          "tipoContrato": {
            "type": "string"
          },
          "dedicacion": {
            "type": "string"
          },
          "activo": {
            "type": "boolean"
          }
        },
        "required": [
          "perfilId"
        ]
      },
      "DocenteResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "perfilId": {
            "type": "string",
            "format": "uuid"
          },
          "codigoDocente": {
            "type": "string"
          },
          "tituloAcademico": {
            "type": "string"
          },
          "departamento": {
            "type": "string"
          },
          "tipoContrato": {
            "type": "string"
          },
          "dedicacion": {
            "type": "string"
          },
          "activo": {
            "type": "boolean"
          },
          "creadoEn": {
            "type": "string",
            "description": "Valor Java OffsetDateTime"
          },
          "actualizadoEn": {
            "type": "string",
            "description": "Valor Java OffsetDateTime"
          },
          "nombres": {
            "type": "string"
          },
          "apellidos": {
            "type": "string"
          }
        }
      },
...
      "PerfilCreateRequest": {
        "type": "object",
        "properties": {
          "identificacion": {
            "type": "string"
          },
          "nombres": {
            "type": "string"
          },
          "apellidos": {
            "type": "string"
          },
          "emailInstitucional": {
            "type": "string"
          },
          "emailPersonal": {
            "type": "string"
          },
          "telefono": {
            "type": "string"
          },
          "direccion": {
            "type": "string"
          },
          "fechaNacimiento": {
            "type": "string",
            "format": "date"
          },
          "fotoUrl": {
            "type": "string"
          }
        },
        "required": [
          "nombres",
          "apellidos",
          "emailInstitucional"
        ]
      },
...
      "PerfilResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "identificacion": {
            "type": "string"
          },
          "nombres": {
            "type": "string"
          },
          "apellidos": {
            "type": "string"
          },
          "emailInstitucional": {
            "type": "string"
          },
          "emailPersonal": {
            "type": "string"
          },
          "telefono": {
            "type": "string"
          },
          "direccion": {
            "type": "string"
          },
          "fechaNacimiento": {
            "type": "string",
            "format": "date"
          },
          "fotoUrl": {
            "type": "string"
          },
          "activo": {
            "type": "boolean"
          },
          "creadoEn": {
            "type": "string",
            "description": "Valor Java OffsetDateTime"
          },
          "actualizadoEn": {
            "type": "string",
            "description": "Valor Java OffsetDateTime"
          }
        }
      },
      "PerfilUpdateRequest": {
        "type": "object",
        "properties": {
          "identificacion": {
            "type": "string"
          },
          "nombres": {
            "type": "string"
          },
          "apellidos": {
            "type": "string"
          },
          "emailInstitucional": {
            "type": "string"
          },
          "emailPersonal": {
            "type": "string"
          },
          "telefono": {
            "type": "string"
          },
          "direccion": {
            "type": "string"
          },
          "fechaNacimiento": {
            "type": "string",
            "format": "date"
          },
          "fotoUrl": {
            "type": "string"
          }
        },
        "required": [
          "nombres",
          "apellidos",
          "emailInstitucional"
        ]
      },
...
      "UsuarioInstitucionalCreateRequest": {
        "type": "object",
        "properties": {
          "perfil": {
            "$ref": "#/components/schemas/PerfilCreateRequest"
          },
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "passwordInicial": {
            "type": "string"
          },
          "rol": {
            "type": "string"
          },
          "pisoId": {
            "type": "string",
            "format": "uuid"
          },
          "carreraId": {
            "type": "string",
            "format": "uuid"
          },
          "periodoId": {
            "type": "string",
            "format": "uuid"
          },
          "nivel": {
            "type": "integer",
            "format": "int32"
          }
        },
        "required": [
          "perfil",
          "username",
          "email",
          "passwordInicial",
          "rol"
        ]
      },
      "UsuarioInstitucionalUpdateRequest": {
        "type": "object",
        "properties": {
          "authId": {
            "type": "string",
            "format": "uuid"
          },
          "perfil": {
            "$ref": "#/components/schemas/PerfilUpdateRequest"
          },
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "rol": {
            "type": "string"
          },
          "activo": {
            "type": "boolean"
          },
          "pisoId": {
            "type": "string",
            "format": "uuid"
          },
          "carreraId": {
            "type": "string",
            "format": "uuid"
          },
          "periodoId": {
            "type": "string",
            "format": "uuid"
          },
          "nivel": {
            "type": "integer",
            "format": "int32"
          }
        },
        "required": [
          "authId",
          "perfil",
          "username",
          "email",
          "rol"
        ]
      }
    }
  }
}